Add a menu option to split assignment weights
It's nice to be able to split weights evenly among all assignments in a course. I'd like to upgrade this in the future at some point so that you can set the weights of certain assignments, the split the remaining ones. Not sure what the UX for that should be yet, though.