Change tasks page to be a calendar
Instead of simple lists of tasks for today and this week, the '/tasks' page now shows a full month calendar. Each task shows up as a colored dot on the day it's due, with a tooltip on hover that displays the name of the associated checklist item (if it exists) or assignment.

The color for each task dot is determined by the 'color' column on the course the task is associated with. This is a new column which was created as part of this calendar view effort.

You can complete or uncomplete a task by clicking on its dot.

My intention with this effort is to make it easier to retroactively or proactively check off tasks. Last quarter, I realized that I don't check the app every day, even if I do have my tasks in mind. I stopped using the app, though, because there was no way to go back and check things off - once I got behind, I felt like the data was already inaccurate, so there was no point in continuing to use it. Another pain point is when I would finish tasks early, I had to wait until the due date arrived in order to check it off. The calendar view fixes all the pain points for me.